dont depend on working network connectivity
In the past, ipset had a dependency to a working network connection
(with available ip addresses and routes and stuff like this), if
systemd-networkd was used. This doesn't make any sense and might even
introduce dependency cycles. We can start as soon as the kernel is ready
to accept calls from ipset. And we need to start before iptables tries
to use us. This will be fixed in a seperate commit.